129 - move walk_puml_graph folder
