+++ tel2puml/pv_to_puml/pv_to_puml.py
@@ -13,15 +13,15 @@ from tel2puml.pv_to_puml.data_ingestion import (
 )
 from tel2puml.pv_event_simulator import transform_dict_into_pv_event
 from tel2puml.events import create_graph_from_events
-from tel2puml.walk_puml_graph.walk_puml_logic_graph import walk_nested_graph
-from tel2puml.walk_puml_graph.node_update import (
+from tel2puml.pv_to_puml.walk_puml_graph.walk_puml_logic_graph import walk_nested_graph
+from tel2puml.pv_to_puml.walk_puml_graph.node_update import (
     update_nested_node_graph_with_break_points,
 )
 from tel2puml.loop_detection.detect_loops import detect_loops
-from tel2puml.walk_puml_graph.create_node_graph_from_event_graph import (
+from tel2puml.pv_to_puml.walk_puml_graph.create_node_graph_from_event_graph import (
     create_node_graph_from_event_graph,
 )
-from tel2puml.walk_puml_graph.find_and_add_loop_kill_paths import (
+from tel2puml.pv_to_puml.walk_puml_graph.find_and_add_loop_kill_paths import (
     find_and_add_loop_kill_paths_to_nested_graphs,
 )
 from tel2puml.puml_graph import (
