+++ tel2puml/pv_to_puml/walk_puml_graph/node_update.py
@@ -3,7 +3,7 @@ from typing import Iterable
 
 from networkx import DiGraph
 
-from tel2puml.walk_puml_graph.node import Node, SubGraphNode
+from tel2puml.pv_to_puml.walk_puml_graph.node import Node, SubGraphNode
 from tel2puml.tel2puml_types import PUMLEvent
