+++ tests/tel2puml/legacy_loop_detection/walk_puml_graph/test_node_legacy.py
@@ -6,7 +6,7 @@ from typing import Literal
 from pm4py import ProcessTree
 import networkx as nx
 
-from tel2puml.walk_puml_graph.node import Node
+from tel2puml.pv_to_puml.walk_puml_graph.node import Node
 from tel2puml.legacy_loop_detection.walk_puml_graph.node import (
     load_logic_tree_into_nodes,
     load_all_logic_trees_into_nodes,
