+++ tests/tel2puml/walk_puml_graph/conftest.py
@@ -9,7 +9,7 @@ from networkx import DiGraph
 
 from tel2puml.events import Event
 from tel2puml.logic_detection import Operator as Logic_operator
-from tel2puml.walk_puml_graph.node import Node, SubGraphNode
+from tel2puml.pv_to_puml.walk_puml_graph.node import Node, SubGraphNode
 from tel2puml.tel2puml_types import (
     DUMMY_END_EVENT,
     DUMMY_START_EVENT,
