+++ tests/tel2puml/walk_puml_graph/test_find_and_add_loop_kill_paths.py
@@ -2,8 +2,8 @@
 
 from networkx import DiGraph
 
-from tel2puml.walk_puml_graph.node import Node, SubGraphNode
-from tel2puml.walk_puml_graph.find_and_add_loop_kill_paths import (
+from tel2puml.pv_to_puml.walk_puml_graph.node import Node, SubGraphNode
+from tel2puml.pv_to_puml.walk_puml_graph.find_and_add_loop_kill_paths import (
     find_and_add_loop_kill_paths_to_nested_graphs,
     find_and_add_loop_kill_paths_to_sub_graph_node,
 )
